+++ .github/workflows/github-actions.yml
@@ -29,10 +29,6 @@ jobs:
 
   mutant:
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        ruby-version: ['2.6', '2.7', '3.0']
-
     env:
       BUNDLE_GEM__MUTANT__DEV: ${{ secrets.MUTANT_KEY }}
     steps:
@@ -45,7 +41,7 @@ jobs:
     # uses: ruby/setup-ruby@v1
       uses: ruby/setup-ruby@473e4d8fe5dd94ee328fdfca9f8c9c7afc9dae5e
       with:
-        ruby-version: ${{ matrix.ruby-version }}
+        ruby-version: '3.0'
         bundler-cache: true # runs 'bundle install' and caches installed gems automatically
     - name: Run Mutant Check
       run: bundle exec mutant run
